Show a clear warning when we're waiting for the DB to become available
The `waitUntilAvailable` set to 30 seconds by default makes the
`connect()` call to retry connecting multiple times, until either the DB
is available ot the timeout is hit. This can create an impression that
"edgedb.connect()" is hung indefinitely. Fix this by prninting a warning
to stderr explaining what is happening.

The warning also suggests to try `connect()` without arguments if the
program is run within a directory created with `edgedb project init` yet
`connect()` was called with explicit arguments.

Lastly, ensure that UNIX sockets are never used for non-Admin
connections.