Support new functional `if..then..else` syntax
We still map this to the old style in the generated EdgeQL for now.
Since the reflection doesn't give us separate operator definitions, we
don't have an easy way to reflect which form is available to you.

We can make this distinction in the future by providing an optional
`__form__` expression property and generating the appropriate EdgeQL
based on that, but for now since we don't have an easy way to tell if
the EdgeDB server supports it without hardcoding some notion of that
into the generated logic, seems best to not do this yet.