Remove duplicate `edgedb` argument
In `npx some-command`, the third `process.argv` argument ends up being
the script name (in this case `"edgedb"`) instead of the first argument
to pass to the binary. Detect if the first argument is `"edgedb"` and
drop it.

Note: If we ever support a subcommand of `edgedb` called `edgedb`, this
might break running the script directly from node, but I don't think
we'll ever introduce such a strange pattern.