Add note for Node 18 flag for crypto global (#952)
Node 18 requires passing a flag to enable the `globalThis.crypto` global
which our PKCE generation code depends on. We used to shim this, but due
to wanting to get away from using `require` directly, decided to drop
the shim. We might reintroduce it, but in the meantime, we should note
this so implementors are not left guessing about this issue.