Test for hard-to-detect recursive CLI calls
Yarn Berry uses an even different scheme for running bin scripts: it
unpacks the script into a temporary directory and adds that temporary
directory into the shell's environment.

This mitigate is by far the most expensive, so it's used dead-last to
avoid the overhead. Looks like on yarn 4.1.1, the call that succeeds
takes an additional 140ms, so that calling latency is added to every
call of the package bin script here.