Check globalThis.Buffer to avoid runtime error
Trying to access `Buffer` directly from global rather than from the
`globalThis` object results in a runtime error on Cloudflare (and likely
other runtimes as well).

Also, it turns out we were never hitting the `Buffer` case because we
were comparing the function to the string `"function"` rather than
comparing the type! And because we were using `tsignore` we were not
getting a helpful type error. Eek.