Look up by operator
Sacrifices the progressive constraint based on the first argument for
much much cheaper inference.

The generation code here has a lot of catching up to do to actually emit
generator-like code, so we need to drop the TypeScript AST
building/printing and adopt the existing code generation tools from the
generator package so it can work with all of the existing tooling around
the typesystem.