Add `BaseType` to queries type
This is true at runtime, so it was failing when I was trying to get
`"unknown"` out of the type since nothing was actually using this.

Also following perf best practices, changed types to interfaces to avoid
needing to do the intersection at type inference time every time it is
being compared.