Update infixContainer* defs to derive RHS from LHS
We know generically that certain types of RHS operands must be derived
from the LHS operand, and it's consistent for all operators. We use this
to build some specific special-casing of how the type unions are built.

Also use `interface` instead of `type` where possible to reduce the
amount of work the type checker has to do.