wip Begin to investigate pointer cardinality issue
If you assign an expression to a property, the returned type does not
properly infer the cardinality from the expression and always just uses
the pointer's cardinality. This is incorrect since we allow assigning
different cardinality if it's a subset of the pointer's cardinality. See
the `cardutil.assignable` helper for the explanation.