Refactor `e.op` better inference perf (#1039)
The existing way of defining the op function simply builds a separate overload
for every operator. That causes the type checker to instantiate a lot of types
depending on how far down the list the overload is defined for that operator and
operands.

This change switches to defining separate overloads based on the operator's
arity (unary, binary, ternary) and the operators return element and cardinality.
That is strictly worse for those operators who were defined pretty close to the
"top", but much better overall. It also seems to be the case that combinators
(like and and or) meant that the number of type instantiations would explode by
some factor (maybe exponential?) depending on how nested the operations got.
This new technique is much more constant.