Use given cardinality in select expression (#1043)
Previously, if you made a select expression with a constant or coalescing
expression, we always calculated the cardinality of a object key based on the
pointer's cardinality, rather than allowing you to return something assignable
to that cardinality. This change will check if the assignment is possible, and
override the cardinality when converting the query builder expression to a TS
type.