Fix weird issue with typechecking after building
For some reason that seems entirely related to environmental issues on
GitHub Action runners, running typechecking after building on the
generate package now fails with this change. I cannot reproduce this
locally, even with a fresh checkout, and after a lot of trial and error,
it seems like the order of running typechecking seems to be the culprit,
but again, only on GHA.

In the end, we do _not_ need to type check these particular packages
because we actual build them, so this we actually useless, but I wanted
to see if I could figure it out before just dropping it.