Use specific type-level marker to `e.shape`
Introduces a type-level marker to the return value of the $shape
function. This marker is used solely for type inference to carry
element, cardinality, and shape information and is not present at
runtime.

We now detect this specific construct in the `setToTsType` (which is the
actual type we alias as `$infer`) and just short-circuit to calculating
the object shape as if the shape was wrapped in a select.