Add support for SNI when connecting to the instance (#1088)
I will try to add these tomorrow:

Documentation is missing:
- SNI,
- Cloud Profile,
- tlsCA maybe should be documented too,
- `credentials` is not documented.

**More things to be done**: add SNI to credentials.json in the cli, and
later read it from there: `Oh and to address this specifically, although
very marginal, we may want to update the CLI to include tls_server_name
in credentials.json if --tls-server-name was specified while linking an
instance.`

Atm CLOUD_PROFILE is taken from env var when using options. And when
using env vars it is not used at all?

---------

Co-authored-by: Scott Trinh <scott@edgedb.com>