Keep verifier cookie in built-in signup
When `require_verification` is false (which is the case here), we should
not delete the PKCE session since we still need it for verifying emails
even though it's not required. We still send the email, and the
verification flow requires that the same PKCE session is valid in the
browser.