Doc fixes from quickstart (#1142)
Had some feedback about some friction in the quickstart flow, so I went
through it myself again and found lots of small things we can do to
improve some of this. There are larger changes we should consider as
well, but this seems like a good low-effort/high-value start. Especially
egregious is that we were using the `Movie` type, but not selecting the
full shape that fits that interface! 😱