Fix querybuilder generated query when using unpathed scoped expr (#1192)
Fixes #808

When walking the expr tree the scoped copy of a select/update expr
passed to the shape should have been treated as an opaque reference and
not walked, since it will get walked again by the enclosing
select/update.

Currently we are only checking this in path expressions (eg.
`movie.title`) and not when the scoped expr appears anywhere else (eg.
`movie.is(e.Movie)`). And since the expr is getting walked multiple
times, it's messing up the ref counting, and exprs are getting hoisted
when they shouldn't, causing the broken queries in #808.